Really remove output files under `output/` directory
It is not enough to just exclude them in .gitignore rules.  They must be
removed from the index so that they are no longer tracked in the future.
The `output/` directory now includes its own .gitignore file, so that
the directory will be created when doing a checkout.  This helps prevent
path-not-found error when running class from the repo directory.